+++ pages/api/space/saveSpace.ts
@@ -9,6 +9,7 @@ import {
 import { DeepPartial } from "typeorm";
 import { v4 as uuidv4 } from "uuid";
 import { SpaceList } from "@entities/SpaceList.entity";
+import { Emps } from "@/server/entities/Emps.entity";
 
 export default async function handler(
   req: AuthenticatedRequest & NextApiRequest,
@@ -51,34 +52,48 @@ export default async function handler(
             where: { UID: userUid },
           });
 
+          // 사용자의 레벨과 보유한 스페이스 수 확인
+          const empsRepository = dataSource.getRepository(Emps);
+
+          const findUser = await empsRepository.findOne({
+            where: { UID: userUid },
+          });
+
+          if (!findUser || !findSpaceList) {
+            return res.status(200).json({
+              message: "사용자 정보를 찾을 수 없습니다.",
+              resultCode: false,
+            });
+          }
           if (findSpaceList) {
-            findSpaceList.space.push(uid);
-            const updateSpaceList = await spaceListRepository.save(
-              findSpaceList
-            );
-
-            if (updateSpaceList) {
+            if (findUser.user_level === 1 && findSpaceList.space.length >= 3) {
               return res.status(200).json({
-                message: "스페이스가 생성되었습니다.",
-                data: { spaceUid: space.UID },
-                resultCode: true,
-              });
-            } else {
-              if (saveSpace) {
-                await spaceRepository.delete(saveSpace.UID);
-              }
-
-              return res.status(200).json({
-                message: "스페이스 생성에 실패하였습니다.",
+                message: "참여할 수 있는 스페이스는 최대 3개입니다.",
                 resultCode: false,
               });
+            } else {
+              findSpaceList.space.push(uid);
+              const updateSpaceList = await spaceListRepository.save(
+                findSpaceList
+              );
+
+              if (updateSpaceList) {
+                return res.status(200).json({
+                  message: "스페이스가 생성되었습니다.",
+                  data: { spaceUid: space.UID },
+                  resultCode: true,
+                });
+              } else {
+                if (saveSpace) {
+                  await spaceRepository.delete(saveSpace.UID);
+                }
+
+                return res.status(200).json({
+                  message: "스페이스 생성에 실패하였습니다.",
+                  resultCode: false,
+                });
+              }
             }
-          } else {
-            return res.status(200).json({
-              message: "스페이스가 생성되었습니다.",
-              data: { spaceUid: space.UID },
-              resultCode: true,
-            });
           }
         } else {
           return res.status(200).json({
